requirements.txt
PySide6==6.5.2
requests~=2.28.1
loguru==0.7.2
tencentcloud-sdk-python-teo==3.0.1019
pyinstaller==6.3.0
packaging~=24.2
tencentcloud-sdk-python~=3.0.1275
cloudflare~=2.8.15